requirements.txt
streamlit
fastapi
uvicorn
pandas
openpyxl
requests
beautifulsoup4
python-dotenv
pydantic
huggingface_hub
pytest
pytest-asyncio
httpx
plotly